# backend/swagger.yaml
swagger: "2.0"
info:
  title: "SCANAR API - System kontroli dostępu"
  description: "API dla systemu weryfikacji dostępu pracowników (QR + rozpoznawanie twarzy)"
  version: "1.0.0"
  contact:
    name: "SCANAR Team"
host: "localhost:5000"
basePath: "/api"
schemes:
  - "http"
...
paths:
...
  /verification/qr:
    post:
      tags:
        - "Verification"
      summary: "Weryfikacja QR kodu"
      description: "Skanuje QR kod i zwraca dane pracownika"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Dane QR kodu do weryfikacji"
          required: true
          schema:
            type: "object"
            required:
              - qr_data
            properties:
              qr_data:
                type: "string"
                example: "a1b2c3d4e5f6..."
                description: "Hash QR kodu pracownika"
      responses:
        200:
          description: "QR kod zweryfikowany poprawnie"
          schema:
            type: "object"
            properties:
              success:
                type: "boolean"
                example: true
              message:
                type: "string"
                example: "Witaj, Jan Kowalski"
              employee:
                type: "object"
                properties:
                  id:
                    type: "integer"
                    example: 1
                  name:
                    type: "string"
                    example: "Jan Kowalski"
                  position:
                    type: "string"
                    example: "Engineer"
                  department:
                    type: "string"
                    example: "IT"
                  qr_code_hash:
                    type: "string"
                  is_active:
                    type: "boolean"
                  created_at:
                    type: "string"
                    format: "date-time"
        400:
          description: "Brak danych QR"
          schema:
            type: "object"
            properties:
              success:
                type: "boolean"
                example: false
              message:
                type: "string"
                example: "Brak danych QR kodu"
              employee:
                type: "null"
        404:
          description: "Nie znaleziono pracownika"
          schema:
            type: "object"
            properties:
              success:
                type: "boolean"
                example: false
              message:
                type: "string"
                example: "Nie znaleziono pracownika dla podanego QR kodu"
              employee:
                type: "null"
        500:
          description: "Błąd serwera"
...
  /verification/employees/{employee_id}/match:
    post:
      tags:
        - "Verification"
      summary: "Porównaj przesłane zdjęcie z encodigami pracownika"
      description: "Przesyła obraz i zwraca czy dopasowano do jednego z zapisanych encodigów. Uses a fixed matching tolerance of 0.6 (not configurable via the API)."
      consumes:
        - "multipart/form-data"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          name: "employee_id"
          required: true
          type: "integer"
          description: "ID pracownika"
        - in: "formData"
          name: "image"
          type: "file"
          required: true
          description: "Plik obrazu z twarzą"
      responses:
        200:
          description: "Zwraca wynik dopasowania"
          schema:
            type: "object"
            properties:
              match:
                type: "boolean"
                example: true
        400:
          description: "Brak obrazu lub brak twarzy w obrazie"
        404:
          description: "Nie znaleziono pracownika"
        500:
          description: "Błąd serwera"
  /verification/save_qr:
    post:
      tags:
        - "Verification"
      summary: "Zapisz wygenerowany kod QR pracownika"
      description: "Odbiera obraz kodu QR w formacie Base64 oraz nazwę pracownika. Dekoduje dane binarne i zapisuje je jako plik .png w dedykowanym folderze na serwerze. Jeśli plik o danej nazwie już istnieje, zostaje nadpisany."
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          required: true
          description: "Obiekt JSON zawierający dane obrazu i identyfikację pracownika"
          schema:
            type: "object"
            properties:
              image:
                type: "string"
                description: "Obraz w formacie Base64 (ciąg znaków zaczynający się od 'data:image/png;base64,')"
                example: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAA..."
              name:
                type: "string"
                description: "Pełne imię i nazwisko pracownika (wykorzystywane jako nazwa pliku)"
                example: "Jan Kowalski"
      responses:
        200:
          description: "Pomyślnie zapisano plik na serwerze"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Saved to ./QRcode/Jan Kowalski.png"
        400:
          description: "Błąd walidacji - brak pola 'image' lub 'name' w żądaniu"
        500:
          description: "Wewnętrzny błąd serwera (np. błąd dekodowania Base64 lub brak uprawnień do zapisu pliku)"
...